+++ inspirer-actix/Cargo.toml
@@ -10,8 +10,9 @@ failure_derive = "0.1.5"
 http = "0.2"
 actix-web = "2.0"
 actix-http = "^1.0"
-serde = "1.0"
-serde_derive = "1.0"
+actix-service = "^1.0"
 serde_json = "1.0"
 log = "0.4.6"
 derive_more = "0.15.0"
+jsonwebtoken = "7"
+serde = {version = "1.0", features = ["derive"] }
